+++ immutable_tree.go
@@ -148,20 +148,27 @@ func (t *ImmutableTree) Has(key []byte) (bool, error) {
 	return t.root.has(t, key)
 }
 
+func (t *ImmutableTree) hashAlgo() HashAlgo {
+	if t.ndb == nil {
+		return HashSHA256
+	}
+	return t.ndb.opts.hashAlgo()
+}
+
 func (t *ImmutableTree) useBlake3() bool {
-	return t.ndb != nil && t.ndb.opts.UseBlake3
+	return t.hashAlgo() == HashBLAKE3
 }
 
 func (t *ImmutableTree) newLeaf(key, value []byte) *Node {
 	n := NewNode(key, value)
-	n.useBlake3 = t.useBlake3()
+	n.algo = t.hashAlgo()
 	return n
 }
 
 // Hash returns the root hash.
 func (t *ImmutableTree) Hash() []byte {
 	if t.root == nil {
-		return emptyDigest(t.useBlake3())
+		return emptyDigest(t.hashAlgo())
 	}
 	return t.root.hashWithCount(t.version + 1)
 }