+++ node.go
@@ -71,7 +71,7 @@ type Node struct {
 	rightNode     *Node
 	subtreeHeight int8
 	isLegacy      bool
-	useBlake3     bool
+	algo          HashAlgo
 }
 
 var _ cache.Node = (*Node)(nil)
@@ -96,10 +96,10 @@ func (node *Node) GetKey() []byte {
 
 // MakeNode constructs an *Node from an encoded byte slice (SHA-256 leaf rehash).
 func MakeNode(nk, buf []byte) (*Node, error) {
-	return makeNode(nk, buf, false)
+	return makeNode(nk, buf, HashSHA256)
 }
 
-func makeNode(nk, buf []byte, useBlake3 bool) (*Node, error) {
+func makeNode(nk, buf []byte, algo HashAlgo) (*Node, error) {
 	// Read node header (height, size, key).
 	height, n, err := encoding.DecodeVarint(buf)
 	if err != nil {
@@ -128,7 +128,7 @@ func makeNode(nk, buf []byte, useBlake3 bool) (*Node, error) {
 		size:          size,
 		nodeKey:       GetNodeKey(nk),
 		key:           key,
-		useBlake3:     useBlake3,
+		algo:          algo,
 	}
 
 	// Read node body.
@@ -330,7 +330,7 @@ func (node *Node) clone(tree *MutableTree) (*Node, error) {
 		rightNodeKey:  node.rightNodeKey,
 		leftNode:      leftNode,
 		rightNode:     rightNode,
-		useBlake3:     node.useBlake3,
+		algo:          node.algo,
 	}, nil
 }
 
@@ -434,13 +434,13 @@ func (node *Node) _hash(version int64) []byte {
 		return node.hash
 	}
 
-	h := getDigest(node.useBlake3)
+	h := getDigest(node.algo)
 	if err := node.writeHashBytes(h, version); err != nil {
-		putDigest(node.useBlake3, h)
+		putDigest(node.algo, h)
 		return nil
 	}
 	node.hash = h.Sum(nil)
-	putDigest(node.useBlake3, h)
+	putDigest(node.algo, h)
 
 	return node.hash
 }
@@ -451,21 +451,21 @@ func (node *Node) _hash(version int64) []byte {
 // to conform with RFC-6962.
 func (node *Node) hashWithCount(version int64) []byte {
 	if node == nil {
-		return emptyDigest(false)
+		return emptyDigest(HashSHA256)
 	}
 	if node.hash != nil {
 		return node.hash
 	}
 
-	h := getDigest(node.useBlake3)
+	h := getDigest(node.algo)
 	if err := node.writeHashBytesRecursively(h, version); err != nil {
-		putDigest(node.useBlake3, h)
+		putDigest(node.algo, h)
 		// writeHashBytesRecursively doesn't return an error unless h.Write does,
 		// and hash.Hash.Write doesn't.
 		panic(err)
 	}
 	node.hash = h.Sum(nil)
-	putDigest(node.useBlake3, h)
+	putDigest(node.algo, h)
 
 	return node.hash
 }
@@ -534,7 +534,7 @@ func (node *Node) writeHashBytes(w io.Writer, version int64) error {
 
 		// Indirection needed to provide proofs without values.
 		// (e.g. ProofLeafNode.ValueHash)
-		valueHash := sum256(node.useBlake3, node.value)
+		valueHash := sum256(node.algo, node.value)
 
 		err = encoding.Encode32BytesHash(w, valueHash[:])
 		if err != nil {